feat: Display message when no audio records are found in audio list page

+++ Controllers/AudioController.php
@@ -45,26 +45,30 @@ class AudioController
         }
         $manager = new Manager();
         $audios = $manager->readTableAll('audio', $userId);
-        $list = "<br><br><dl class='dl'>";
-        foreach ($audios as $audio) {
-            $id = $audio['id'];
-            $title = htmlspecialchars($audio['title'], ENT_QUOTES, 'UTF-8');
-            $artist = htmlspecialchars($audio['artist'], ENT_QUOTES, 'UTF-8');
-            $image = htmlspecialchars($audio['image'], ENT_QUOTES, 'UTF-8');
-            $afficher = "<input type='checkbox' class='select-audio' data-audio-id='$id'>";
-            $modifier = "<a class='btnBase vert' href='index.php?url=audio/update/$id'>Modifier</a>";
-            $supprimer = "<a class='btnBase rouge' href='index.php?url=audio/delete/$id'>Supprimer</a>";
-            $list .= "<div class='audio-item'>";
+        if (empty($audios)) {
+            $list = "<p>Aucun enregistrement audio trouvé.</p>";
+        } else {
+            $list = "<br><br><dl class='dl'>";
+            foreach ($audios as $audio) {
+                $id = $audio['id'];
+                $title = htmlspecialchars($audio['title'], ENT_QUOTES, 'UTF-8');
+                $artist = htmlspecialchars($audio['artist'], ENT_QUOTES, 'UTF-8');
+                $image = htmlspecialchars($audio['image'], ENT_QUOTES, 'UTF-8');
+                $afficher = "<input type='checkbox' class='select-audio' data-audio-id='$id'>";
+                $modifier = "<a class='btnBase vert' href='index.php?url=audio/update/$id'>Modifier</a>";
+                $supprimer = "<a class='btnBase rouge' href='index.php?url=audio/delete/$id'>Supprimer</a>";
+                $list .= "<div class='audio-item'>";
 
-            $list .= "<img class='photoAudio' src='Ressources/images/pochettes/$image' alt='cover'>";
-            $list .= "<div class='audio-details'>";
-            $list .= "<h4>$title</h4>";
-            $list .= "<p>$artist</p>";
+                $list .= "<img class='photoAudio' src='Ressources/images/pochettes/$image' alt='cover'>";
+                $list .= "<div class='audio-details'>";
+                $list .= "<h4>$title</h4>";
+                $list .= "<p>$artist</p>";
 
-            $list .= "<div class='button-group'>$afficher $modifier $supprimer</div>";
-            $list .= "</div></div>";
+                $list .= "<div class='button-group'>$afficher $modifier $supprimer</div>";
+                $list .= "</div></div>";
+            }
+            $list .= "</dl><br><br>";
         }
-        $list .= "</dl><br><br>";
         return $list;
     }
